% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{add_numbers}
\alias{add_numbers}
\title{Add two numbers}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{Another number.}
}
\value{
The sum of x and y.
}
\description{
This function takes two numbers and returns their sum.
}
