% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4010Project1.R
\name{RP}
\alias{RP}
\title{Power Function}
\usage{
RP(x, power)
}
\arguments{
\item{x}{Must be a numeric object.}

\item{power}{A numeric object called \code{power}.}
}
\value{
Raises values in numeric object \code{x} to value of \code{power}.
}
\description{
Power Function
}
\examples{
RP(c(1,3,5), 3)
RP(x = c(1,3,5), power = 3)
RP(power = 2, x = c(1, 3, 5))
}

